See description:
Have the energy bar working but currently it is attached to only the RMB for testing purposes and points per click is on the player. Will change to abilities once added to game.

Fixed bug where a duplicate game canvas was instantiating at runtime and messing up our energy bar.